+++ src/resource/lifecycle/index.ts
@@ -1,5 +1,10 @@
 // Copyright 2026 PeopleWare N.V.
 // SPDX-License-Identifier: Apache-2.0
 export * from "./audited.schema.js";
+export * from "./auditable-persistent.schema.js";
+export * from "./auditable.schema.js";
 export * from "./created-in-error.schema.js";
+export * from "./insert-auditable-persistent.schema.js";
+export * from "./insert-auditable.schema.js";
+export * from "./persistent.schema.js";
 export * from "./timestamped.schema.js";